Update deploy.sh file

+++ stop.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
+if pgrep current/helpdesk.jar
+
+then
+
 kill $(ps aux | grep '[c]urrent/helpdesk.jar' | awk '{print $2}')
+
+fi
